+++ docs/main/changelog.rst
@@ -10,7 +10,7 @@ follows `Semantic versioning`_
 Development version
 -------------------
 - Update documentation and rework examples for: ``Singleton``, ``Callable``, ``Coroutine``,
-  ``Object``, ``List``, ``Configuration``, ``Selector`` providers.
+  ``Object``, ``List``, ``Configuration``, ``Selector``, and ``Dependency`` providers.
 - Fix mypy stub of the ``DeclarativeContainer`` to specify the ``__init__`` interface.
 
 3.34.0